requirements.txt
APScheduler==3.5.3
Flask==1.0.2
mysql-connector==2.1.6
mysql-connector-python==8.0.13
mysql-connector-python-rf==2.2.2
requests==2.21.0
xmljson==0.2.0
xmltodict==0.11.0
bcrypt==3.1.5
Flask-Cors==3.0.7
python-dateutil==2.7.5
hdfs==2.2.2
matplotlib==3.0.2
python-docx==0.8.10
